third_party/blink/web_tests/external/wpt/webtransport/datagrams.https.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] Fail to transfer max-size+1 datagram, handle PMTUD increases
  assert_greater_than: expected a number greater than 1024 but got 1024
[FAIL] Datagram MaxAge getters/setters work correctly
  assert_equals: expected (number) Infinity but got (object) null
[FAIL] Datagram HighWaterMark getters/setters work correctly
  assert_throws_js: function "() => { wt.datagrams.incomingHighWaterMark = -1; }" did not throw
Harness: the test ran to completion.
